extract provider interfaces to separate file
